// src/app/modules/academics/grading/models/class-assesment-grades-param-list-view-model.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


import {ClassAssesmentGradesParamViewModel} from './class-assessment-grades-param-view-model';

/**
 * ClassAssesmentGradesParamListViewModel
 */
export interface ClassAssesmentGradesParamListViewModel { 
    /**
     * Gets Or Sets ClassAssesmentGradesParamViewModel
     */
    classAssesmentGradesParamViewModel?: Array<ClassAssesmentGradesParamViewModel>;
}