// src/app/modules/academics/grading/models/class-assessmnet-garde-filter-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * 
 */
export interface ClassAssessmentFilterView {
    sortBy: any;
    schoolBoardId?: any;
    classId?: any;
    sectionId?: any;
    gradingTermId?: any;
    assessmentCategoryId?: any;
    assessmentGroupId?: any;
    assessmentParameterId?: any;
    sortOrder: 0 | 1 | -1;
    schoolAcademicSessionStudentId: any;
}